// src/app/models/vida-y-salud/EstadoLoteVidaSalud.ts
// CARGADO -> VALIDANDO -> VALIDADO_OK|VALIDADO_CON_ERRORES|ERROR_VALIDACION -> PROCESANDO_DOCUMENTOS
// (transitorio, fase 3) -> FACTURANDO -> FACTURADO|FACTURADO_CON_ERRORES|ERROR_FACTURACION
export enum EstadoLoteVidaSalud {
  CARGADO = 'CARGADO',
  VALIDANDO = 'VALIDANDO',
  VALIDADO_OK = 'VALIDADO_OK',
  VALIDADO_CON_ERRORES = 'VALIDADO_CON_ERRORES',
  ERROR_VALIDACION = 'ERROR_VALIDACION',
  PROCESANDO_DOCUMENTOS = 'PROCESANDO_DOCUMENTOS',
  FACTURANDO = 'FACTURANDO',
  FACTURADO = 'FACTURADO',
  FACTURADO_CON_ERRORES = 'FACTURADO_CON_ERRORES',
  ERROR_FACTURACION = 'ERROR_FACTURACION',
  ELIMINADO = 'ELIMINADO',
}